For Issue 46 - added a file config/config.yml with site specific attributes. Did a first pass at moving some constants out of the application and into this file.
Never edit config.yml to make local changes, instead copy anything you want to modify localy to config/local.yml and change them there. (these directions are in the config.yml file). Since local.yml is not in subversion, you can modify it as much as you want. Please be careful not to add local.yml to subversion by accident.

*Do* edit config.yml to put as many constants as you think should go there. Anything that would change if the application were running on a different server. Anything that would change if the application were being used by another group for their own archive.

Anything in config.yml (or local.yml) is available to the application as AppConfig.yaml_label