issue 71 - FixtureReplacement
This is the coolest thing! 

I'm checking it in, because it doesn't break any existing tests that use the old fixtures. All I've done so far is to define example_data.rb to such an e
xtent that create_model does not error for all the models listed. They may not make sense, but they don't error.

I'll start actually writing some tests that use this plugin next.

In the meantime, if you're interested, you can use

$ script/console
>> include FixtureReplacement
=> Object
>> create_creatorship
=> #<Creatorship id: 1, creation_id: 1, creation_type: "Chapter", pseud_id: 1, created_at: "2008-05-16 02:21:21", updated_at: "2008-05-16 02:21:21">
>>
>> Pseud.find(1)
=> [#<Pseud id: 1, user_id: 1, name: "I Cwrjvno evq", description: "A pr etqski", is_default: false, created_at: "2008-05-16 02:21:21", updated_at: "2008-05-16 02:21:21">]
and watch your database fill up