BROKEN COMMIT, DO NOT DEPLOY! Because I've touched so many files at this point, I want to get this chunk of work committed to lessen the danger of overwriting, but these revisions are not yet fully working.
The comment system has been redesigned to match the wireframes, and refactored to simplify and to improve degradation in non-javascript-compatible browsers. I've opted for doing this using new names, getting the new stuff working, then deleting the old, rather than editing the existing, to make it easier to catch any places where references to the old ones are hiding using file search. 

What should currently be working, with and without javascript:
- showing/hiding comments inline on chapters and on works
- recursive comment threading to a limited depth
- opening and closing inline reply forms both to chapters/works and to other comments

What is not yet working:
- submitting/deleting new comments

Won't Fix issue:
If you first use javascript to open up the comments and then turn off Javascript, the non-javascript fallback links end up with the wrong code, and weird results ensue. See the note in helpers/comments_helper.rb for more details. Someone consistently browsing without javascript should NOT have problems, and a simple reload of the original page will fix it, so I think this is tolerable. 

Other edits: 
- work navigation (first/next/prev/last chapter) is now in. 
- bookmark link will no longer appear on works you have already bookmarked (ideally we should have a "delete" link in its stead)
- minor refactoring of application layout and works views to reduce file size
- placeholder icon

Stuff this breaks: 
- only comments-related tests are failing (as expected), but I have not yet done a full pass over all the views/partials I've touched to make sure they won't affect anything else. Also, the inbox will not be working until the comments are fixed.

Whew! Sorry for the monster commit --